Restructured repo to utilise modules
