+++ Dockerfile_user_code
@@ -15,7 +15,9 @@ COPY requirements.txt .env /opt/dagster/app
 
 RUN pip install -r requirements.txt
 
-COPY pipeline.py fpl_getter.py gcs_parquet_io_manager.py /opt/dagster/app
+COPY /fpl_dagster .env /opt/dagster/app/fpl_dagster
+
+WORKDIR /opt/dagster/app/fpl_dagster
 
 # Run dagster gRPC server on port 4000
 
@@ -25,6 +27,6 @@ EXPOSE 4000
 # HEALTHCHECK CMD curl --fail http://localhost:4000
 # CMD allows this to be overridden from run launchers or executors that want
 # to run other commands against your repository
-CMD ["dagster", "api", "grpc", "-h", "0.0.0.0", "-p", "4000", "-f", "pipeline.py"]
+CMD ["dagster", "api", "grpc", "-h", "0.0.0.0", "-p", "4000", "-f", "definitions.py"]
 
 #"--inject-env-vars-from-instance"
